% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadMod.R
\name{Bstep}
\alias{Bstep}
\title{Bstep
Function to increment haddock biomass 1 year}
\usage{
Bstep(B, qE = 0.15, sdU = 0.1, dt = 0.1, fB = c("SH", "schaefer"), ...)
}
\arguments{
\item{B}{biomass in kt}

\item{qE}{catchability*effort = F (harvest rate); F*B = C, catch}

\item{sdU}{standard deviation of the random shock; set to 0 for deterministic simulation}

\item{dt}{within-season size of time step; should be a number > 0 and <= 1}

\item{fB}{function of B; should be a character of either "SH" or "schaefer", corresponding to one of the two models}

\item{...}{arguments passed to fB}
}
\description{
Bstep
Function to increment haddock biomass 1 year
}
\details{
Follows equation 1 in Spencer and Collie 1997.
}
\note{
In the example below, the equilibrium value (average between 2020 and 2040) for F=0.18 is different under low-noise and high-noise scenarios. Below the simulation is as in the paper, but if the noise level is decreased (or if simulated deterministically), then the equiblirum is much closer to the cases of F=0.21 and F=0.24. I think this is a noise-induced transition. Cool!
}
\examples{

# Make Figure 5a
nY <- 50
# 13.8 kt is from pg 2922, second column, penultimate paragraph; 
# 13.8 is the starting value in 1993
makeB <- function(){c(13.8, rep(NA, nY-1))}
Year <- seq(1993, length.out=nY)
qE <- c(0, 0.06, 0.12, 0.18, 0.21, 0.24)
Bvec <- replicate(length(qE), makeB())
for(j in 1:length(qE)){
	for(i in 2:nrow(Bvec)){
		Bvec[i,j] <- Bstep(Bvec[i-1,j], qE[j], sdU=0.1)
	}
}
ltys <- c("solid", "dotted", "dashed", "longdash", "twodash", "dotdash")
ylim <- range(Bvec, na.rm=TRUE)
plot(Year, Bvec[,1], type='l', 
	lty=ltys[1], ylim=ylim, 
	ylab="Biomass (kt)", xlab="time", main="STH model"
)
for(j in 2:ncol(Bvec)){
	lines(Year, Bvec[,j], lty=ltys[j])
}
legend("topleft", paste("F", qE, sep="="), lty=ltys)

}

