% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redShift.R
\name{plot.rslist}
\alias{plot.rslist}
\title{Plot redShift List
Plot a list of outputs from redShift}
\usage{
\method{plot}{rslist}(x, ...)
}
\arguments{
\item{x}{a list of outputs from \code{\link{redShift}}}

\item{...}{parameters passed to graphics::image}
}
\description{
Plot redShift List
Plot a list of outputs from redShift
}

