% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redShift.R
\name{print.rs}
\alias{print.rs}
\title{Print rs
Print red shift output w/o the spec and freq}
\usage{
\method{print}{rs}(x, ...)
}
\arguments{
\item{x}{output of \code{\link{redShift}}}

\item{...}{to be consistent with print}
}
\description{
Print rs
Print red shift output w/o the spec and freq
}

