% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redShift.R
\name{redShift}
\alias{redShift}
\title{Red Shift
Function to calculate the spectral slope; warning signal known as "red shift"}
\usage{
redShift(x, ...)
}
\arguments{
\item{x}{numeric vector that is a regular (evently spaced) time series}

\item{...}{not used}
}
\value{
spectral slope
}
\description{
Red Shift
Function to calculate the spectral slope; warning signal known as "red shift"
}

