% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollWarn.R
\name{rollWarn}
\alias{rollWarn}
\title{Roll Warn
Calculate early warnings statistics over a rolling window}
\usage{
rollWarn(x, stat = eval(formals(ews)$stat), win = 15)
}
\arguments{
\item{x}{numeric vector over which to calculate statistics; should be a regularly spaced time series}

\item{stat}{name of statistic to use; see \code{\link{ews}}}

\item{win}{window length}
}
\value{
numeric vector of a rolling window statistic
}
\description{
Roll Warn
Calculate early warnings statistics over a rolling window
}
\examples{
# ==========================================
# = Set up simulation and analysis options =
# ==========================================
nYear <- 5E2
Year <- 1:nYear
tripVal <- c(0.36, 0.21)
nBurn <- 50
sdU <- 0.05
qE_range <- c(tripVal[1]-0.2, tripVal[1]+0.1)
seq_arg <- c(as.list(qE_range), list(length.out=nYear))
qE <- do.call(seq, seq_arg)

# ==============================
# = Claculate time series of B =
# ==============================
B0 <- 400
B0 <- Burn(B0, n=nBurn, qE=qE[1], sdU=0)
Bvec <- c(B0, rep(NA, nYear-1))
for(i in 2:nYear){
	Bvec[i] <- Bstep(B=Bvec[i-1], qE=qE[i], sdU=sdU)
}

# ==================================
# = Calculate rolling window stats =
# ==================================
sdVec <- rollWarn(Bvec, stat='sd', win=min(50, nYear/5))
ac1Vec <- rollWarn(Bvec, stat='ac1', win=min(50, nYear/5))
redVec <- rollWarn(Bvec, stat="redShift", win=min(50, nYear/5))

# =======================
# = Plot B and Warnings =
# =======================
dev.new(width=6, height=6)
par(mfrow=c(2,2), mar=c(1.75,1.75,0.5,1.75), ps=8, cex=1, mgp=c(0.75,0.15,0), tcl=-0.15)
plot(Year, Bvec, type='l')
qECrit2 <- Year[which.min(abs(qE-tripVal[1]))]
abline(v=qECrit2, lty='dashed')
text(qECrit2, y=1*max(Bvec), label=paste0("qE = ", tripVal[1]), pos=2)

plot(Year, sdVec, col="forestgreen", type="l", ylab="Standard Deviation")
abline(v=qECrit2, lty='dashed')
plot(Year,ac1Vec, col="blue", type='l', ylab="AR(1)")
abline(v=qECrit2, lty='dashed')
plot(Year, redVec, col="red", type='l', ylab="redShift")
abline(v=qECrit2, lty='dashed')
}
\seealso{
\code{\link{ews}}
}

