% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadMod.R
\name{schaefer}
\alias{schaefer}
\title{Schaefer Model
Schaefer Model for Haddock biomass change. Default parameter values are for haddock biomass in Spencer and COllie (1997)}
\usage{
schaefer(B, r = 0.4, K = 129, ...)
}
\arguments{
\item{B}{biomass of haddock in previous time step}

\item{r}{intrinsic growth rate of haddock}

\item{K}{carrying capacity of haddock}

\item{...}{not used}
}
\description{
Schaefer Model
Schaefer Model for Haddock biomass change. Default parameter values are for haddock biomass in Spencer and COllie (1997)
}
\details{
Fit is from 1976-1993, which authors reported as being "reasonable" (there was difficulty in getting a good fit to data from this model, particularly compared to Steele and Henderson model)
}
\seealso{
\code{\link{SH}} \code{\link{Bstep}}
}

