% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadMod.R
\name{SH}
\alias{SH}
\title{Steele and Henderson Model
Steele and Henderson model for haddock biomass. Default parameters are empirically fitted values from Spencer and Collie (1997).}
\usage{
SH(B, r = 0.76, K = 520.2, D = 28.2, A = 27.9, ...)
}
\arguments{
\item{B}{biomass of haddock in previous time step}

\item{r}{intrinsic growth rate of haddock}

\item{K}{carrying capacity of haddock}

\item{D}{maximum rate of predation in type III predator functional response}

\item{A}{predator saturation in type III predator functional response}
}
\description{
Steele and Henderson Model
Steele and Henderson model for haddock biomass. Default parameters are empirically fitted values from Spencer and Collie (1997).
}
\details{
The default arguments are the parameter values from Table 1 of Spencer and Collie 1997
}
\seealso{
\code{\link{schaefer}} \code{\link{Bstep}}
}

